SQL_Analysis/requirements.txt
# Python Dependencies
# Install with: pip install -r requirements.txt

pandas>=1.3.0
mysql-connector-python>=8.0.0
scikit-learn>=1.0.0

